Revert change of radius to i4 from i8 - causes crash.
Turns out using i4 with a radius of 7 you can actually overflow the
integer *because of the sign bit*. So u4 and i4 can hold very different
numbers, with i4 holding *half* of u4 because of the sign bit.

If I just use i8, I'm totally fine. i5 would work as well, but I'm
honestly not gaining anything here.