Optimize TemporalSoften Vec, scale thresholds based on bit depth.
There's actually a few changes bundled up here, including faster
min/max's for vectors, courtesy of zmath examples, and improvements
in the loadVec and storeVec methods since we can derive type and length
details from one parameter instead of needing two.

In terms of optimization for TemporalSoften, it really comes down to the
fact that I was using way wider bit depths for my intermediary casts
than was necessary. This was confused when I was originally building the
sum type, which requires wider bit depths on purpose.

The scalar version of Temporal Soften still needs the same cleanup.