Add RG mode 2, and possible vec version of Mode 1.
The Zig optimizer is pretty darn good, and my custom vector version
isn't any faster than the auto-vectorized scalar implementation, and may
actually be slower.

I'm going to prioritize writing the scalar versions for now, which
really seem to scream and are 2-6x faster than the RGVS and RGSF
versions.