Add shared sort function for RG Mode 5+
Also cleaned up signed type determination so that it's determined once
at the top of the RG struct and shared everywhere (that needs it). It
still might need some tweaking, like calculating f16 as f32 for speed,
or f32 as f64 for accuracy, but I'm leaving that for a later
exercise.