Implemented threading using a thread pool, and optional work group.
Basically, I tried creating a Thread Pool in Get Frame,
and creating a threadpool in Create.

Creating a shared threadpool in Create seemed about the fastest.
of around 480 fps compared to 450 fps for the Get Frame implementation.

Using a threadpool in Get Frame instead of just raw threads was about
the same, maybe slightly slower than raw threads (450 vs 460 fps).

So a shared thread pool, created in Create is the fastest,
but it's still not faster than no threads at all.