Fix segfaults in Median/Soften.
Turns out my src frame freeing refactor was buggy.

Essentially, because the src_frame array is of a known max size,
but we can a number of frames *less* than this max size, I was
making extra free calls to pointers that I hadn't initialized.

Since we know the diameter or frame count at runtime, we just use that,
as its always accurate.