Optimized the scalar version of FluxsmoothT
Now it's almost as fast, if not as fast as my own vectorized version.

There's some cool integer division tricks used, but using floating point
arithmetic is just that much faster.

Need to do the same thing in the spatial/temporal version as well.

Completely removes the need for the weird rounding magic numbers...