Adjust stride calculation in Temporal Median.
Other filters still need to be fixed.

The previous calculation was unnecessarily complex and completely
ignored the actual vector size when calculating an offset. This is
extremely error prone and showed up on users systems with strides that
range up to 64 bytes (instead of my machine's 32-byte stride).

This new calculation is similar to my prior solution from several
releases ago, but properly includes stride when calculating the height
of the current row.

Hopefully

Fixes: https://github.com/adworacz/zsmooth/issues/1