Major speedup of Temporal Median.
By using comptime, I completely remove the branching that previously
happened in the switch in medianVector. I'm now making the diameter
known at comptime and then switching to use the proper call to
processPlane at runtime using an switch with an inline prong.

This brings the number of branches down substantially, and in my tests
takes the TemporalMedian from ~305fps to 370 fps.

At this point, 2/3rds of our performance cost is just loading/storing
memory, so there's not much more we can do.