Change benchmarks to use YUV420 instead of RGB.
YUV420 is the much more common format by far, so
it's much more likely to align with user's expectations when they start
using Zsmooth after seeing these benchmarks.

I remember going with RGB because it was the only one that had a
half-float (f16) format in Vapoursynth at the time, but that was
addressed for YUV several VS versions ago.

But since YUV is faster (it's less data, since the U and V planes take
up half the size of the Y plane, whereas RGB is the same size for all
planes), any previously stated performance numbers are now moot.

So I'll have to rerun benchmarks again...

Also, I bumped up the base number of frames for running benchmarks to
2000, which I might bring up even higher if both my 9950X and M4 Mac see
similar performance numbers. My 9950X is freakin' fast on this kind of
content... I'm definitely getting into the realm of being memory
bandwidth constrained on a number of filters.

For now I'm just using the `--frame-count-scale` param in my benchmark
to bump things up on the fly.