Add benchmark entries for Zig 0.14.0 (AVX2 / AVX512).
There's a big gain in performance for most (but not all) filters using
AVX512, with some regressions (looking at you, RemoveGrain mode 1/4, and
a few others). But on average we're looking at a ~30% improvement, with
some filters going ~94% improvement (DegrainMedian).

I also tested the difference between znver4 and znver5, and znver5 pulls
out ahead *verrrryyy* slightly, but it's mostly within margin of error,
so running znver4 on znver5 CPUs gives 99.9% of the speed benefit.

This gap may widen once LLVM actually merges Zen 5 tunings, which will
then require Zig to upgrade to said LLVM version, and then release a new
version... so it might be a while at this rate (which is a little sad,
given how long Zen 5 has been out).

But for now, AVX2 builds and znver4 are great offerings.

Also of note, upgrading to Zig 0.14.0 did see a *slight* speed
reduction, but again we're in the realm of margin of error. My
calculations showed a ~0.69% (yes, that's .69 *percent*) decrease of
overall filter performance going from Zig 0.12.1 to Zig 0.14.0 for AVX2.

But performance was better overall for Zig 0.14.0 with AVX512 (znver4),
sometimes significantly so (RemoveGrain 1 and 4 saw a nice boost from
0.12.1 to 0.14.0, but not enough to make up the performance
delta with the (still faster) AVX2 version of RemoveGrain).

So I'm calling the Zig 0.14.0 upgrade "good enough" at this point.