Redo benchmarks for AVX512.
Previous benchmarks weren't sorted alphabetically, didn't include
TTempSmooth, and didn't include other plugins as a point of comparison
(only the AVX2 results had other plugins).

This is a full redo, and is much more fair as I manually recompiled most
if not all of the other plugins, leading to massive increases in
performance for things like the older RemoveGrain plugin.