Move format resolution to getFrame for Median.
This pattern should be applied to all of my filters, but it does several
things for such a small change:
1. It shrinks the output binary size (not by a whole lot, maybe ~1-3KB
   per filter), but hey, why not.
2. It self-contains each filter kernel, which:
3. Ensures maximum portability (so we can easily add support for
   Avisynth or FFMPEG in the future)
4. Ensures maximum testability, since each filter kernel just takes in
   byte slices and parameters, meaning we can easily unit test/fuzz
   test.
5. Surprisingly, leads to some performance improvements in some filters.

I don't entirely understand this last one. It could be that LLVM is
better able to layout code for optimal instruction caching / alignment.

It may be able to conduct some other optimizations because we no longer
bubble up filters to the `getFrame` functions and we keep the filter
kernel completely self-contained.

But I don't really know, and I'm not going to look the gift horse in the
mouth on this one.