Fix release builds so that they don't clobber native builds.
Previously, we were overwriting the root_module configuration in the
release target loop, even when we weren't actually conducting a release
build. This caused some wonky behavior, including building for the wrong
architectures / OSes.

We now create a separate copy of the root module and library
configuration for each target.