simplify median function signature
I was overly aggressive in forcing comptime evaluation for said
function. Just using a slice is enough, as the compiler is smart enough
to recognize when the slice length is comptime known.

I ran benchmarks for the Median, SmartMedian, and TemporalMedian
functions, and they're basically identical in performance. Only minor
variations (which might have more to do with my BIOS update + not adding
the 100Mhz overclock).